Second revision
Quite a dirty way of doing this. This script was made by me a couple years ago and I will need to clean this up quite a bit now that I know better. 

This extension added 64bit capability along with shotgunning nasty .exe's to try to get rid of them.